preliminary float matrix

+++ TOMB4/game/delstuff.h
@@ -9,8 +9,8 @@ void Rich_CalcLaraMatrices_Interpolated(short* frame1, short* frame2, long frac,
 void CalcLaraMatrices(long flag);
 
 extern short* GLaraShadowframe;
-extern long lara_matrices[180];
-extern long lara_joint_matrices[180];
+extern float lara_matrices[180];
+extern float lara_joint_matrices[180];
 extern long LaraNodeAmbient[2];
 extern long bLaraUnderWater;
 extern uchar LaraNodeUnderwater[15];
